+++ app/src/main/java/kr/ksw/mybudget/presentation/add/card/screen/AddCardScreen.kt
@@ -47,7 +47,7 @@ import kr.ksw.mybudget.data.local.entity.CARD_TYPE_CREDIT
 import kr.ksw.mybudget.data.local.entity.CARD_TYPE_DEBIT
 import kr.ksw.mybudget.presentation.add.common.components.ADD_INPUT_TYPE_CARD
 import kr.ksw.mybudget.presentation.add.common.components.AddInputForm
-import kr.ksw.mybudget.presentation.add.transformation.CardNumberTransformation
+import kr.ksw.mybudget.presentation.add.common.transformation.CardNumberTransformation
 import kr.ksw.mybudget.presentation.components.CardUi
 import kr.ksw.mybudget.presentation.core.common.Paddings
 import kr.ksw.mybudget.presentation.core.common.getCardColor
